refactor: Memo でまとめた
