+++ core/supabase/src/main/java/app/kaito_dogi/mybrary/core/supabase/memo/MemoResponse.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.supabase.model
+package app.kaito_dogi.mybrary.core.supabase.memo
 
 import app.kaito_dogi.mybrary.core.data.dto.MemoDto
 import kotlinx.serialization.SerialName
